minor change to design
